sources: refactor script source into module.
Extract the Script source into `sources/_script.py`.

LP: #1648906
